requirements.txt
docling
fastapi
uvicorn
pypdf
PyMuPDF
pillow
google-generativeai
python-dotenv
python-multipart
openai
